fix: Đổi lại database id là Bigint
- Fix các trường dữ liệu Id thành Long(thay vì Integer ở 1 vài class)
- Sửa liên quan ở JpaRepo về Long
- Sửa dto id về Long

+++ MedicalAppointmentBookingSystem/src/main/java/org/mabs/repository/MedicineRepository.java
@@ -3,6 +3,6 @@ package org.mabs.repository;
 import org.mabs.entity.Medicine;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MedicineRepository extends JpaRepository<Medicine, Integer> {
+public interface MedicineRepository extends JpaRepository<Medicine, Long> {
 
 }
